implemented new logging scheme based on asymm_array_2d
Note: ended up making a change to originate_particle_simplex.  It was
putting the new particle at iFound (rather than i_particle) regardless of
its status relative to target.  That probably does not matter, since a point
outside of target would have resulted in a call to _shift_particle, anyway,
which would accomplish what was supposed to be accomplished.  I have just moved
the work to a different point in the code.